+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added configuration for hiding/showing document regions on model and canvas page
 - Added new Annotation styles for customising colors and borders of annotations
 - Added new zoom controls
+- Enabled new project template hook: `beforeCloneModel`
+- Enabled new project template hook: `onRevisionApproved`
 
 ### Fixed
 - Regression with dropdown (reverts fix for overflow bug in popup)